+++ Editor/BuildingBlocks/BuildingBlockElements.cs
@@ -9,7 +9,7 @@ namespace Unity.XR.CoreUtils.Editor.BuildingBlocks
     /// in the Scene View allowing the user to easily access and drop these elements in the current scene.
     /// Implement this interface to define a Building Block. In order to add this block to the UI, this Building Block
     /// should either be returned by an <see cref="IBuildingBlockSection.GetBuildingBlocks"/> enumerable (common case),
-    /// or else identified as an orphan Building Block by directly using the <see cref="BuildingBlockItemAttribute"/>
+    /// or else identified as an unsectioned Building Block by directly using the <see cref="BuildingBlockItemAttribute"/>
     /// on the Building Block class.
     /// </summary>
     public interface IBuildingBlock
@@ -23,6 +23,16 @@ namespace Unity.XR.CoreUtils.Editor.BuildingBlocks
         /// The path to the icon of this Building Block in the UI. This icon should be placed in a Resources folder.
         /// </summary>
         string IconPath { get; }
+
+        /// <summary>
+        /// Whether the user know if this Building Block is enabled or disabled. If disabled, the Building Block will be grayed out in the UI.
+        /// </summary>
+        bool IsEnabled { get; }
+
+        /// <summary>
+        /// Description of the Building Block. This description will be displayed as a tooltip in the UI.
+        /// </summary>
+        string Tooltip { get; }
 
         /// <summary>
         /// Implement this method to execute the actions associated to this Building Block.
@@ -59,9 +69,9 @@ namespace Unity.XR.CoreUtils.Editor.BuildingBlocks
 
     /// <summary>
     /// Attribute used to register a class as a Building Block Section or a Building Block.
-    /// Using this attribute directly on a Building Block will register the class in the manager as a orphan
+    /// Using this attribute directly on a Building Block will register the class in the manager as an unsectioned
     /// Building Block.
-    /// Orphan building blocks which will display blocks outside sections in the interface. This is not recommended since
+    /// Unsectioned building blocks which will display blocks outside sections in the interface. This is not recommended since
     /// it makes it prone to confusion with other packages.
     /// The target class must derive from either <see cref="IBuildingBlockSection"/> or <see cref="IBuildingBlock"/>.
     /// </summary>
@@ -70,7 +80,7 @@ namespace Unity.XR.CoreUtils.Editor.BuildingBlocks
     {
         /// <summary>
         /// The priority of this item used to order elements in the overlay.
-        /// Orphans Building Blocks will be ordered and displayed together first, then all the Building Block Sections
+        /// Unsectioned Building Blocks will be ordered and displayed together first, then all the Building Block Sections
         /// will follow and be ordered between them.
         /// </summary>
         public int Priority = 0;